producer-weather/requirements.txt
pymongo
python-dotenv
asyncio
requests
boto3
aiohttp